pyproject.toml
[project]
name = "neuromem"
version = "0.7.0"
description = "Memory management framework for AI agents"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Jacky", email = "jacky@example.com"}
]
keywords = ["ai", "memory", "agent", "llm", "rag", "vector-database"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",
    "httpx>=0.27.0",
]

[project.urls]
Homepage = "https://github.com/MatrixDriver/neuromem"
Documentation = "https://github.com/MatrixDriver/neuromem"
Repository = "https://github.com/MatrixDriver/neuromem"
Issues = "https://github.com/MatrixDriver/neuromem/issues"

[project.optional-dependencies]
s3 = ["boto3>=1.34.0"]
pdf = ["pypdf>=4.0.0"]
docx = ["python-docx>=1.1.0"]
all = [
    "boto3>=1.34.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
]
local-embedding = ["sentence-transformers>=3.0.0"]
eval = ["tqdm>=4.66.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "boto3>=1.34.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["neuromem*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "slow: tests that require external API (embedding/LLM)",
    "requires_db: tests that require PostgreSQL",
]
